commander cody, execute order 66 Removed OpenGG Test, made all math classes immutable, major changes to resource loading, part 1 of texture remake, made a bunch of stuff safer, deleted in-jar resources, moved all resources to OpenGG

+++ OpenGG Project/src/com.opengg.core/classes/com/opengg/core/world/components/SunComponent.java
@@ -95,9 +95,9 @@ public class SunComponent extends RenderComponent{
     @Override
     public void serialize(GGOutputStream stream) throws IOException{
         super.serialize(stream);
-        stream.write((int)SUN_DIS);
-        stream.write((float)rotspeed);
-        stream.write((float)currot);
+        stream.write(SUN_DIS);
+        stream.write(rotspeed);
+        stream.write(currot);
         stream.write(texture.getData().get(0).source);
     }
